+++ android/app/src/main/kotlin/deckers/thibault/aves/MainActivity.kt
@@ -13,6 +13,7 @@ import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import android.os.TransactionTooLargeException
+import android.os.ext.SdkExtensions
 import android.provider.MediaStore
 import android.util.Log
 import androidx.annotation.RequiresApi
@@ -371,7 +372,7 @@ open class MainActivity : FlutterFragmentActivity() {
                     if (action == MediaStore.ACTION_REVIEW_SECURE) {
                         val uris = ArrayList<String>()
                         intent.clipData?.let { clipData ->
-                            for (i in 0 until clipData.itemCount) {
+                            for (i in 0..<clipData.itemCount) {
                                 clipData.getItemAt(i).uri?.let { uris.add(it.toString()) }
                             }
                         }
@@ -399,12 +400,34 @@ open class MainActivity : FlutterFragmentActivity() {
                 }
             }
 
-            Intent.ACTION_GET_CONTENT, Intent.ACTION_PICK -> {
+            Intent.ACTION_GET_CONTENT,
+            Intent.ACTION_PICK,
+            MediaStore.ACTION_PICK_IMAGES -> {
+                // common intent extras
+                var allowMultiple = intent.getBooleanExtra(Intent.EXTRA_ALLOW_MULTIPLE, false)
+                val mimeTypes = intent.getStringArrayExtra(Intent.EXTRA_MIME_TYPES)?.toList()
+                val pickLocalOnly = intent.getBooleanExtra(Intent.EXTRA_LOCAL_ONLY, false)
+
+                // MediaStore picker intent extras
+                var pickImagesMax = 0
+                var pickInOrder = false
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                    if (SdkExtensions.getExtensionVersion(Build.VERSION_CODES.R) >= 2) {
+                        pickImagesMax = intent.getIntExtra(MediaStore.EXTRA_PICK_IMAGES_MAX, 0)
+                    }
+                    if (SdkExtensions.getExtensionVersion(Build.VERSION_CODES.R) >= 12) {
+                        pickInOrder = intent.getBooleanExtra(MediaStore.EXTRA_PICK_IMAGES_IN_ORDER, false)
+                    }
+                }
+                allowMultiple = allowMultiple || pickImagesMax > 0
+
                 return hashMapOf(
                     INTENT_DATA_KEY_ACTION to INTENT_ACTION_PICK_ITEMS,
                     INTENT_DATA_KEY_MIME_TYPE to intent.type,
-                    INTENT_DATA_KEY_MIME_TYPES to intent.getStringArrayExtra(Intent.EXTRA_MIME_TYPES)?.toList(),
-                    INTENT_DATA_KEY_ALLOW_MULTIPLE to intent.getBooleanExtra(Intent.EXTRA_ALLOW_MULTIPLE, false),
+                    INTENT_DATA_KEY_MIME_TYPES to mimeTypes,
+                    INTENT_DATA_KEY_ALLOW_MULTIPLE to allowMultiple,
+                    INTENT_DATA_KEY_PICK_IN_ORDER to pickInOrder,
+                    INTENT_DATA_KEY_PICK_LOCAL_ONLY to pickLocalOnly,
                 )
             }
 
@@ -605,6 +628,8 @@ open class MainActivity : FlutterFragmentActivity() {
         const val INTENT_DATA_KEY_MIME_TYPE = "mimeType"
         const val INTENT_DATA_KEY_MIME_TYPES = "mimeTypes"
         const val INTENT_DATA_KEY_PAGE = "page"
+        const val INTENT_DATA_KEY_PICK_IN_ORDER = "pickInOrder"
+        const val INTENT_DATA_KEY_PICK_LOCAL_ONLY = "pickLocalOnly"
         const val INTENT_DATA_KEY_QUERY = "query"
         const val INTENT_DATA_KEY_SECURE_URIS = "secureUris"
         const val INTENT_DATA_KEY_URI = "uri"