+++ lib/model/settings/settings.dart
@@ -71,6 +71,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
     _unregister();
     _register(monitorPlatformSettings);
     initAppSettings();
+    await sanitize();
   }
 
   void _unregister() {
@@ -160,6 +161,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
     bottomNavigationActions = [];
     showOverlayOnOpening = false;
     showOverlayMinimap = false;
+    showOverlayZoomLevel = false;
     showOverlayThumbnailPreview = false;
     viewerGestureSideTapNext = false;
     viewerUseCutout = true;
@@ -183,6 +185,11 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
       set(SettingKeys.videoBackgroundModeKey, null);
     }
     collectionBurstPatterns = collectionBurstPatterns.where(BurstPatterns.options.contains).toList();
+
+    // emulator
+    if (!device.isPhysicalDevice) {
+      set(SettingKeys.videoHardwareAccelerationKey, VideoHardwareAcceleration.disabled);
+    }
   }
 
   // tag editor
@@ -373,6 +380,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
             case SettingKeys.tagSortReverseKey:
             case SettingKeys.showOverlayOnOpeningKey:
             case SettingKeys.showOverlayMinimapKey:
+            case SettingKeys.showOverlayZoomLevelKey:
             case SettingKeys.showOverlayInfoKey:
             case SettingKeys.showOverlayDescriptionKey:
             case SettingKeys.showOverlayRatingTagsKey:
@@ -381,7 +389,6 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
             case SettingKeys.viewerGestureSideTapNextKey:
             case SettingKeys.viewerUseCutoutKey:
             case SettingKeys.enableMotionPhotoAutoPlayKey:
-            case SettingKeys.enableVideoHardwareAccelerationKey:
             case SettingKeys.videoGestureDoubleTapTogglePlayKey:
             case SettingKeys.videoGestureSideDoubleTapSeekKey:
             case SettingKeys.videoGestureVerticalDragBrightnessVolumeKey:
@@ -424,6 +431,7 @@ class Settings with ChangeNotifier, SettingsAccess, SearchSettings, AppSettings,
             case SettingKeys.imageBackgroundKey:
             case SettingKeys.videoAutoPlayModeKey:
             case SettingKeys.videoBackgroundModeKey:
+            case SettingKeys.videoHardwareAccelerationKey:
             case SettingKeys.videoLoopModeKey:
             case SettingKeys.videoResumptionModeKey:
             case SettingKeys.subtitleTextAlignmentKey: